feat: added v4.7.1 upgrade handler and removed old handlers
