--- app/upgrades/v450/upgrade.go
@@ -1,46 +1,0 @@
-package v450
-
-import (
-	storetypes "github.com/cosmos/cosmos-sdk/store/types"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/module"
-	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
-
-	"github.com/desmos-labs/desmos/v4/app/upgrades"
-)
-
-var (
-	_ upgrades.Upgrade = &Upgrade{}
-)
-
-// Upgrade represents the v4.5.0 upgrade
-type Upgrade struct {
-	mm           *module.Manager
-	configurator module.Configurator
-}
-
-// NewUpgrade returns a new Upgrade instance
-func NewUpgrade(mm *module.Manager, configurator module.Configurator) *Upgrade {
-	return &Upgrade{
-		mm:           mm,
-		configurator: configurator,
-	}
-}
-
-// Name implements upgrades.Upgrade
-func (u *Upgrade) Name() string {
-	return "v4.5.0"
-}
-
-// Handler implements upgrades.Upgrade
-func (u *Upgrade) Handler() upgradetypes.UpgradeHandler {
-	return func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
-		// Do nothing here as we don't have anything particular in this update
-		return u.mm.RunMigrations(ctx, u.configurator, fromVM)
-	}
-}
-
-// StoreUpgrades implements upgrades.Upgrade
-func (u *Upgrade) StoreUpgrades() *storetypes.StoreUpgrades {
-	return &storetypes.StoreUpgrades{}
-}
